bugfix and optimization

+++ Arduino/Quadricottero2.0/Quadricottero2.0.ino
@@ -2,9 +2,9 @@
 #include <EEPROM.h>
 #include <PCintPulseIn.h>
 #include <Servo.h>
-#include "I2Cdev.h"
-#include "MPU6050_6Axis_MotionApps20.h"
-#include "Wire.h"
+#include <I2Cdev.h>
+#include <MPU6050_6Axis_MotionApps20.h>
+#include <Wire.h>
 
 //#######RX######
 
@@ -12,7 +12,7 @@
 * always sorted this way:
 * THROTTLE, PITCH, ROLL, YAW, POT_A
 */
-const byte RXpins[] = {A0, A1, A2, A3, 8};
+const int RXpins[] = {A0, A1, A2, A3, 8};
 const unsigned int signalMin[] = {1172, 1176, 1160, 1132, 2008};
 const unsigned int signalMax[] = {1816, 1816, 1944, 1940, 996};
 int throttle;
@@ -100,3 +100,4 @@ byte flightMode = ACRO_MODE;
 
 byte showLog = false;
 //#define LOG_LOOP_SPEED
+//#define LOG_FREE_RAM
